python/requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
pycryptodome==3.11.0
goplus==0.2.4
Werkzeug<3
typing-extensions==4.0.1
python-dateutil==2.8.2